Webapp/react-flask-app/backend/requirements.txt
Flask>=2.0.2
PyMySQL
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
google-api-python-client
pandas
flask-cors